try fp16 SDPA tensor-core path (smaller bias, fused kernel)

+++ autoresearch/submission/submission.py
@@ -121,15 +121,20 @@ def block_sparse_attn_fwd(q, k, v, row_ptr, col_idx, seq_lens):
         # Build the float bias in ONE write pass via torch.where (vs
         # torch.zeros + masked_fill_, which is two passes over the ~2 GB
         # tensor on full-suite shapes).
+        # Build bias in fp16 directly (4× smaller alloc than fp32) and feed
+        # fp16 q/k/v to SDPA so the fused tensor-core mem-efficient kernel
+        # runs instead of the much slower fp32 fallback. fp16 has 11
+        # mantissa bits → ~5e-4 quantization noise on the output, well
+        # under the 1e-3 atol. LSE is still returned in fp32.
         attn_bias = torch.where(
             invalid_4d,
-            torch.tensor(neg_inf, device=device, dtype=torch.float32),
-            torch.tensor(0.0, device=device, dtype=torch.float32),
+            torch.tensor(neg_inf, device=device, dtype=torch.float16),
+            torch.tensor(0.0, device=device, dtype=torch.float16),
         )
         B = batch_heads * nq
-        q_for_sdpa = q_chunks.reshape(B, 1, BLOCK_SIZE, head_dim)
-        k_for_sdpa = gathered_k.reshape(B, 1, md * BLOCK_SIZE, head_dim)
-        v_for_sdpa = gathered_v.reshape(B, 1, md * BLOCK_SIZE, head_dim)
+        q_for_sdpa = q_chunks.reshape(B, 1, BLOCK_SIZE, head_dim).to(torch.float16)
+        k_for_sdpa = gathered_k.reshape(B, 1, md * BLOCK_SIZE, head_dim).to(torch.float16)
+        v_for_sdpa = gathered_v.reshape(B, 1, md * BLOCK_SIZE, head_dim).to(torch.float16)
         bias_for_sdpa = attn_bias.reshape(B, 1, BLOCK_SIZE, md * BLOCK_SIZE)
         out_sdpa, lse_sdpa, _seed, _offset = torch.ops.aten._scaled_dot_product_efficient_attention(
             q_for_sdpa,
@@ -141,7 +146,7 @@ def block_sparse_attn_fwd(q, k, v, row_ptr, col_idx, seq_lens):
             False,  # is_causal
             scale=SCALE,
         )
-        out_blocks = out_sdpa.reshape(batch_heads, nq, BLOCK_SIZE, head_dim)
+        out_blocks = out_sdpa.to(torch.float32).reshape(batch_heads, nq, BLOCK_SIZE, head_dim)
         lse_blocks = lse_sdpa.reshape(batch_heads, nq, BLOCK_SIZE)
         # Zero / -inf invalid query rows (SDPA may leak garbage into them).
         out_blocks = torch.where(
